requirements/backend_requirements.txt
fastapi==0.81.0
jupyter==1.0.0                    # dev
nltk==3.7                         # dev
openpyxl==3.0.10                  # dev
pandas==1.3.5                     # dev
passlib[bcrypt]==1.7.4
pre-commit==2.19.0                # dev
psycopg2-binary==2.9.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
requests==2.28.1                  # dev
scikit-learn==1.0.2               # dev
selenium==4.4.3                   # dev
SQLAlchemy==1.4.40
uvicorn[standard]==0.18.3
